Add files via upload
This is the code for the do while loop. I used an array to get three inputs from user and the program will find the sum of the input from the array. Then finally the program will ask the user to press 1 to continue or any other key to quite program